Switch to standard english
'of recent' works, but it threw me for a bit. Thought it was a grammatical error, came here to fix it, then realised it was 'of recent'. It's not technically wrong, but it might throw any American readers for a loop, so I'd suggest changing it. Figured there was no harm filing a PR about it.